requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
drf-spectacular==0.18.1
importlib-metadata==4.8.1
inflection==0.5.1
jsonschema==3.2.0
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
zipp==3.5.0
